move LinearInfoForDirectEncoding
